cleanup: fix handling serial number resets
When a new RRDP session starts, the serial number resets. Since the expiration
of a serial was based on having any _higher_ serial that was not expired, any
snapshots in a new session are wiped until they catch up with the old session's
serial. This causes the rsync symlink to point to a non-existing directory,
breaking the repository.

With this patch the serial numbers are only used to track publication timestamps
in state. A serial can only expire when there is a _newer_ publication in
state. I.e. regardless of the age of the latest publication, it will never
expire.